Renamed files for clarify and preventing conflicts
